+++ src/components/motors/TB67H450.h
@@ -5,14 +5,15 @@
 
 class TB67H450 : public MotorDriver {
 private:
-  int in1, in2;                           // PWM 信号用のピン番号
-  void run(int pwmValue1, int pwmValue2); // モーターを制御する内部メソッド
+  uint8_t in1, in2; // PWM 信号用のピン番号
+  void run(uint8_t pwmValue1,
+           uint8_t pwmValue2); // モーターを制御する内部メソッド
 
 public:
-  TB67H450(int in1, int in2); // コンストラクタ、ピン番号を初期化
-  void forward() override;    // モーターを前進させるメソッド
-  void reverse() override;    // モーターを後退させるメソッド
-  void stop() override;       // モーターを停止させるメソッド
+  TB67H450(uint8_t in1, uint8_t in2); // コンストラクタ、ピン番号を初期化
+  void forward() override; // モーターを前進させるメソッド
+  void reverse() override; // モーターを後退させるメソッド
+  void stop() override;    // モーターを停止させるメソッド
   void setSpeed(float rate) override; // モーターの速度を設定するメソッド
 };
