+++ src/utils/DebugLogger.h
@@ -11,7 +11,8 @@ public:
   // シリアルポートを初期化する
   static void init(HardwareSerial &serial, unsigned long baudrate = 19200);
   static void init(SoftwareSerial &serial, unsigned long baudrate = 19200);
-  static void init(int rxPin, int txPin, unsigned long baudrate = 19200);
+  static void init(uint8_t rxPin, uint8_t txPin,
+                   unsigned long baudrate = 19200);
 
   // デバッグメッセージを出力する
   // クラス名、メソッド名、メッセージを指定して出力する
